LPS-71611 Moves test to trash-test osgi module
