--- portal-impl/test/integration/com/liferay/portal/kernel/trash/bundle/trashhandlerregistryutil/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Description: Test Bundle for TrashHandlerRegistryUtilTest
-Bundle-Version: 1.0.0
-Private-Package: $(bundle.package).*
--dsannotations: *
